+++ ntex-net/src/compio/io.rs
@@ -89,13 +89,17 @@ trait Terminate {
 
 impl Terminate for compio_net::TcpStream {
     fn terminate(&self) -> io::Result<()> {
-        socket2::SockRef::from(self).shutdown(std::net::Shutdown::Both)
+        let sock = socket2::SockRef::from(self);
+        crate::helpers::drain_socket(&sock);
+        sock.shutdown(std::net::Shutdown::Both)
     }
 }
 
 impl Terminate for compio_net::UnixStream {
     fn terminate(&self) -> io::Result<()> {
-        socket2::SockRef::from(self).shutdown(std::net::Shutdown::Both)
+        let sock = socket2::SockRef::from(self);
+        crate::helpers::drain_socket(&sock);
+        sock.shutdown(std::net::Shutdown::Both)
     }
 }